fix project logging in this branch

+++ server/data/schema.js
@@ -60,8 +60,8 @@ const typeDefs = `
   }
 
   type Location{
-    lat: Int
-    lng: Int
+    lat: Float
+    lng: Float
   }
 
   type Reputation {
